requirements_cpu.txt
tqdm
requests
googledrivedownloader
opencv-python
scipy
tensorflow-cpu
protobuf
numpy
pillow
